pyproject.toml
[project]
name = "Agile"
version = "0.1.0"
description = "Agile bu ..."
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "alembic>=1.16.4",
    "argon2-cffi>=25.1.0",
    "celery>=5.5.3",
    "fastapi[all]>=0.116.1",
    "flower>=2.0.1",
    "passlib[cryptography]>=1.7.4",
    "pre-commit>=4.2.0",
    "psycopg2-binary>=2.9.10",
    "python-dotenv>=1.1.1",
    "python-jose>=3.5.0",
    "redis>=6.2.0",
    "ruff>=0.12.3",
    "sqlalchemy>=2.0.41",
    "starlette-admin>=0.15.1",
]

# RUFF configs
[tool.ruff]
target-version = "py312"
line-length = 88

# Enable rules
lint.select = ["E", "F", "I", "N", "UP", "B", "C90", "RUF"]

lint.ignore = [
    "E501",     # line too long
    "RUF001",   # ambiguous unicode character in code
    "RUF002",   # ambiguous unicode character in docstring
]  

# Paths to exclude
exclude = ["migrations", "__pycache__", ".venv", "alembic"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
line-ending = "lf"
